tree/: Drop const from opaque data ptr in master definitions [alt]
The opaque data pointer need not necessarily have constant
data for the life-time of the specific master. This is because
the data field purpose is for the master to use as it sees fit
for managing its own internal state and therefore we should not
constrain this as being RO data at init time.

flashrom-stable:
* Applied bitbang_spi part, now that we can test it.
* Added OOM check.
* Use new shutdown API right away.

Change-Id: I6c3cf561313de1de4e3caec8732137a8c23bf1e6
Signed-off-by: Edward O'Callaghan <quasisec@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/flashrom/+/49131
Original-Reviewed-by: Sam McNally <sammc@google.com>
Reviewed-on: https://review.coreboot.org/c/flashrom-stable/+/73266
Reviewed-by: Nico Huber <nico.h@gmx.de>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>