Makefile: Add `versioninfo` target, compare during export
There are other ways to generate a tarball beside our `tarball` make
target. For instance, web-based Git UIs offer to export the tree as
well. Alas, such exports would lack any version information.

A compromise is to check the version information in, but only on
release branches. This is the first commit to add helpers to main-
tain a checked-in `versioninfo.inc` on release branches.

To add a safety net, let the `_export` target fail if the checked-
in `versioninfo.inc` is not up-to-date.

Change-Id: I73cab7346ee470485c5f65ac499aff9cd15d98c7
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/flashrom-stable/+/71778
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>