wbsio_spi: Request memory mapping locally
Our SPI chip drivers are not aware of any memory mapping. However, two
of our SPI master drivers that implement an LPC-to-SPI bridge make use
of the memory mapping originally intended for memory mapped flash chips.

This unnecessarily breaks the abstraction in core flashprog code. There
is no other component than the master driver involved with the mapping,
hence we can simply perform it locally in this driver.

The driver already assumed a maximum memory mapped space of 1MiB. We
always try to map this whole space and than adapt the address offsets
accordingly.

Change-Id: If95bad2ca40611452e34a865d520e9aa88e8f163
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.sourcearcade.org/c/flashprog/+/72520