chipset_enable: Optionally check PCI revision field
We used to match compatible chipset devices by vendor and device ID
only. On some chipsets, e.g. AMD southbridges / SoCs, this is not
enough, though, as the device IDs are rarely updated.

In the case of AMD chipsets, we can identify the chipset with the
revision ID of the SMBus device. So we add that field to the chipset
enable list.

Change-Id: I4021cf8e83c605fde4360c274b39481b1e0ff070
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/flashrom-stable/+/72573
Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>