manibuilder/alpine: Fix SSL cert issues
Once more, try to find a way to build all the containers. They enabled a
redirection from http to https recently and there is no option to skip
certificate checks. So the only move is to provide a current CA certifi-
cate by ourselves. Note, the added certificate is the current Root CA
of Let's Encrypt. Due to issues with the ca-certificates packages in
Alpine v3.7, nothing using another Root CA will work this way.

Change-Id: I3bb340428b863ea5e24ad6e1da6f3fe428eb187d
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/flashrom-stable/+/73402
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>