manibuilder/multiarch: Use persistence feature of qemu-user-static
We can use a single container now to provide `qemu-user` binaries for
multi-arch containers. This allows us to use official multi-arch
containers that don't contain such a binary for the host architecture.

We pin the version to 7.2.0-1 for reproducibility. However, in this
version, `qemu-arm-static` has some trouble running Alpine containers
(programs crashing on 3.14 and 3.15). So keep that one at 6.3.0-8.

Change-Id: Iadf7ac318f99e521ce8c0acc0c41f59711e7c397
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/flashrom-stable/+/73432
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>