test_build: Update make call
Use ccache and parallel builds by default, unless ${CC} or ${CPUS} is
overridden. Also allow ${WARNERROR} to be set to `no` (it defaults to
`yes` in the Makefile already).

The make command can be overridden by setting ${MAKECMD}. This can be
used in testing environments where GNU make is installed as `gmake`,
for instance.

Run `make clean` ahead to make sure that we always have a clean build.

Change-Id: Ie68bbc530474d30aad997c372a2b2f9144aae012
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/flashrom-stable/+/73572
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>