internal: Don't try linux_mtd on x86
We assume that the first MTD device found represents the "internal"
firmware flash. This is true on many architectures, and assumed to
be working. On x86, however, there is traditionally no MTD device.
One exception is the `spi-intel` driver, but this one is tagged
"DANGEROUS" and often makes trouble if a Linux distribution enabled
it nevertheless.

So, let's disable the internal/MTD automatism on x86. Flashrom has
better drivers, and if somebody runs into a situation where the MTD
driver would work but the internal one doesn't, they can still use
`linux_mtd` explicitly.

Change-Id: I813980786a09fe64f541906e1963b0abd8b93cb5
Signed-off-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://review.coreboot.org/c/flashrom-stable/+/73987
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>